--- a/.gitignore
+++ b/.gitignore
@@ -25,3 +25,5 @@ _testmain.go
 
 # until I get it working
 cmd/api-router-gen
+*.bench
+*.txt