--- a/resp.go
+++ b/resp.go
@@ -18,7 +18,7 @@ var (
 	RespBadRequest       Response = NewJSONErrorResponse(http.StatusBadRequest)
 	RespEmpty            Response = &simpleResp{code: http.StatusNoContent}
 	RespOK               Response = &simpleResp{code: http.StatusOK}
-	RespRedirectRoot     Response = Redirect("/", false)
+	RespRedirectRoot              = Redirect("/", false)
 
 	// Break can be returned from a handler to break a handler chain.
 	// It doesn't write anything to the connection.