```sh
━➤ benchstat v2.bench  v4.bench
name                  old time/op    new time/op    delta
JumpRouter5Params-16     132ns ± 2%      79ns ± 1%   -40.42%  (p=0.008 n=5+5)
JumpRouterStatic-16     22.7ns ± 2%    12.9ns ± 0%   -43.27%  (p=0.008 n=5+5)

name                  old alloc/op   new alloc/op   delta
JumpRouter5Params-16     32.0B ± 0%      0.0B       -100.00%  (p=0.008 n=5+5)
JumpRouterStatic-16      0.00B          0.00B           ~     (all equal)

name                  old allocs/op  new allocs/op  delta
JumpRouter5Params-16      1.00 ± 0%      0.00       -100.00%  (p=0.008 n=5+5)
JumpRouterStatic-16       0.00           0.00           ~     (all equal)
```